requirements.txt
streamlit
stability-sdk
Pillow
torch
transformers